fix(cookie-consent): address category registry review feedback
Summary:
- Move category preference key aliasing into a shared helper used by
  cookie persistence, consent logging, and Tracks props.
- Coalesce legacy Tracks preference props to booleans when callers pass
  partial preference objects, and cover that behavior in tests.
- Remove stale consent-type filter setup from the PHP test and drop the
  package prefix from the changelog entry.

Rationale:
- The alias mapping affects multiple registry consumers and should have a
  single implementation to avoid drift.
- The legacy Tracks event schema should keep stable boolean props even when
  dynamic preference maps omit a default key.
- Tests should describe the registry-based allow-list now that the old
  consent-type filter has been removed from the controller path.

Review comments:
- 3490609617
- 3490609686
- 3490609704
- 3490609738

Tests:
- GNUPGHOME=/tmp/symphony-gnupg MISE_DATA_DIR=/tmp/symphony-mise-data
  MISE_CACHE_DIR=/tmp/symphony-mise-cache mise x node@24.15.0 -- pnpm
  --filter @automattic/jetpack-cookie-consent test -- --runTestsByPath
  tests/utils.test.ts tests/tracks.test.ts
- GNUPGHOME=/tmp/symphony-gnupg MISE_DATA_DIR=/tmp/symphony-mise-data
  MISE_CACHE_DIR=/tmp/symphony-mise-cache mise x node@24.15.0 -- pnpm
  --filter @automattic/jetpack-cookie-consent typecheck
- GNUPGHOME=/tmp/symphony-gnupg MISE_DATA_DIR=/tmp/symphony-mise-data
  MISE_CACHE_DIR=/tmp/symphony-mise-cache mise x node@24.15.0 -- pnpm
  --filter @automattic/jetpack-cookie-consent build-production
- composer --working-dir=projects/packages/cookie-consent phpunit --
  --filter Consent_Log_Controller_Test
- php -l projects/packages/cookie-consent/tests/php/Consent_Log_Controller_Test.php
- git diff --check
