+++ projects/packages/cookie-consent/changelog/wooa7s-1599-category-registry
@@ -1,4 +1,4 @@
 Significance: patch
 Type: changed
 
-Cookie Consent: Add a configurable consent category registry.
+Add a configurable consent category registry.
